requirements.txt
python-telegram-bot==12.0.0
simplesqlitewrap==0.1.14
toml
